+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ### 0.10.0 (In progress)
 - Added aarch64 (ARM64) CPU architecture support
 - Added ARM64 .deb, .rpm and .flatpak build support for CI/CD
+- Added transcoding queue
 - Added confirmation dialog when transcoding would overwrite existing file
 - Timeline control handles are now focusable
 - Timeline control handles accept left/right arrow, and h/l keypresses for more granular control
+- Improved visibility of a text popup in Profile Editor
+- Standardized pop-up's
 - Swapped "Copy" and "Ok" buttons in about dialog
+- Added confirmation dialog when pressing "Copy" button in about dialog
 - Unified margins, paddings and spacings across the entire app
 - Fixed encoder compatibility dialog rendering at the bottom on first launch in some Qt Quick Styles such as Fusion